move runner stuff out of tiktorch.server
