+++ tiktorch/runner/prediction_pipeline/_model_adapters/_tensorflow_model_adapter.py
@@ -4,8 +4,7 @@ import numpy as np
 import tensorflow as tf
 import xarray as xr
 from bioimageio.spec import nodes
-from bioimageio.spec.utils import get_nn_instance
-
+from tiktorch.runner.utils import get_nn_instance_from_source
 from ._model_adapter import ModelAdapter
 
 
@@ -19,7 +18,7 @@ class TensorflowModelAdapter(ModelAdapter):
         # FIXME: TF probably uses different axis names
         self._internal_output_axes = _output.axes
 
-        self.model = get_nn_instance(bioimageio_model)
+        self.model = get_nn_instance_from_source(bioimageio_model)
         self.devices = []
         tf_model = tf.keras.models.load_model(spec.weights["tensorflow_saved_model_bundle"].source)
         self.model.set_model(tf_model)
